requirements.txt
alembic==1.14.0
aniso8601==9.0.1
blinker==1.9.0
click==8.1.7
Flask==3.0.3
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
packaging==24.2
passlib==1.7.4
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
